Refactor hole filling into free function
