+++ src/pmp/algorithms/hole_filling.cpp
@@ -1,7 +1,7 @@
 // Copyright 2011-2020 the Polygon Mesh Processing Library developers.
 // Distributed under a MIT-style license, see LICENSE.txt for details.
 
-#include "pmp/algorithms/HoleFilling.h"
+#include "pmp/algorithms/hole_filling.h"
 
 #include <Eigen/Dense>
 #include <Eigen/Sparse>
@@ -12,6 +12,91 @@ using SparseMatrix = Eigen::SparseMatrix<double>;
 using Triplet = Eigen::Triplet<double>;
 
 namespace pmp {
+
+class HoleFilling
+{
+public:
+    HoleFilling(SurfaceMesh& mesh);
+    void fill_hole(Halfedge h);
+
+private:
+    struct Weight
+    {
+        Weight(Scalar _angle = std::numeric_limits<Scalar>::max(),
+               Scalar _area = std::numeric_limits<Scalar>::max())
+            : angle(_angle), area(_area)
+        {
+        }
+
+        Weight operator+(const Weight& _rhs) const
+        {
+            return Weight(std::max(angle, _rhs.angle), area + _rhs.area);
+        }
+
+        bool operator<(const Weight& _rhs) const
+        {
+            return (angle < _rhs.angle ||
+                    (angle == _rhs.angle && area < _rhs.area));
+        }
+
+        Scalar angle;
+        Scalar area;
+    };
+
+    // compute optimal triangulation of hole
+    // throws InvalidInputException in case of a non-manifold hole.
+    void triangulate_hole(Halfedge h);
+
+    // compute the weight of the triangle (i,j,k).
+    Weight compute_weight(int i, int j, int k) const;
+
+    // refine triangulation (isotropic remeshing)
+    void refine();
+    void split_long_edges(const Scalar lmax);
+    void collapse_short_edges(const Scalar lmin);
+    void flip_edges();
+    void relaxation();
+    void fairing();
+
+    // return i'th vertex of hole
+    Vertex hole_vertex(unsigned int i) const
+    {
+        assert(i < hole_.size());
+        return mesh_.to_vertex(hole_[i]);
+    }
+
+    // return vertex opposite edge (i-1,i)
+    Vertex opposite_vertex(unsigned int i) const
+    {
+        assert(i < hole_.size());
+        return mesh_.to_vertex(
+            mesh_.next_halfedge(mesh_.opposite_halfedge(hole_[i])));
+    }
+
+    // does interior edge (_a,_b) exist already?
+    bool is_interior_edge(Vertex _a, Vertex _b) const;
+
+    // triangle area
+    Scalar compute_area(Vertex _a, Vertex _b, Vertex _c) const;
+
+    // triangle normal
+    Point compute_normal(Vertex _a, Vertex _b, Vertex _c) const;
+
+    // dihedral angle
+    Scalar compute_angle(const Point& _n1, const Point& _n2) const;
+
+    // mesh and properties
+    SurfaceMesh& mesh_;
+    VertexProperty<Point> points_;
+    VertexProperty<bool> vlocked_;
+    EdgeProperty<bool> elocked_;
+
+    std::vector<Halfedge> hole_;
+
+    // data for computing optimal triangulation
+    std::vector<std::vector<Weight>> weight_;
+    std::vector<std::vector<int>> index_;
+};
 
 HoleFilling::HoleFilling(SurfaceMesh& _mesh) : mesh_(_mesh)
 {
@@ -473,6 +558,11 @@ void HoleFilling::fairing()
 
     // clean up
     mesh_.remove_vertex_property(vsel);
+}
+
+void fill_hole(SurfaceMesh& mesh, Halfedge h)
+{
+    HoleFilling(mesh).fill_hole(h);
 }
 
 } // namespace pmp
